datateck/requirements.txt
asgiref==3.2.5
bson==0.5.8
dataclasses==0.6
Django==2.2.11
djongo==1.3.1
dnspython==1.16.0
fdb==2.0.1
future==0.18.2
jsonfield==3.1.0
numpy==1.18.2
pandas==1.0.3
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
